+++ src/main/java/com/structurizr/dsl/DeploymentEnvironmentDslContext.java
@@ -1,10 +1,16 @@
 package com.structurizr.dsl;
 
-final class DeploymentEnvironmentDslContext extends DslContext {
+final class DeploymentEnvironmentDslContext extends GroupableDslContext {
 
-    private String environment;
+    private final String environment;
 
     DeploymentEnvironmentDslContext(String environment) {
+        super(null);
+        this.environment = environment;
+    }
+
+    DeploymentEnvironmentDslContext(String environment, ElementGroup group) {
+        super(group);
         this.environment = environment;
     }
 
@@ -15,6 +21,7 @@ final class DeploymentEnvironmentDslContext extends DslContext {
     @Override
     protected String[] getPermittedTokens() {
         return new String[] {
+                StructurizrDslTokens.GROUP_TOKEN,
                 StructurizrDslTokens.DEPLOYMENT_GROUP_TOKEN,
                 StructurizrDslTokens.DEPLOYMENT_NODE_TOKEN,
                 StructurizrDslTokens.RELATIONSHIP_TOKEN