+++ src/main/java/com/structurizr/dsl/DeploymentNodeDslContext.java
@@ -1,13 +1,20 @@
 package com.structurizr.dsl;
 
 import com.structurizr.model.DeploymentNode;
+import com.structurizr.model.GroupableElement;
 import com.structurizr.model.ModelItem;
 
-final class DeploymentNodeDslContext extends ModelItemDslContext {
+final class DeploymentNodeDslContext extends GroupableElementDslContext {
 
-    private DeploymentNode deploymentNode;
+    private final DeploymentNode deploymentNode;
 
     DeploymentNodeDslContext(DeploymentNode deploymentNode) {
+        this(deploymentNode, null);
+    }
+
+    public DeploymentNodeDslContext(DeploymentNode deploymentNode, ElementGroup group) {
+        super(group);
+
         this.deploymentNode = deploymentNode;
     }
 
@@ -21,8 +28,14 @@ final class DeploymentNodeDslContext extends ModelItemDslContext {
     }
 
     @Override
+    GroupableElement getElement() {
+        return deploymentNode;
+    }
+
+    @Override
     protected String[] getPermittedTokens() {
         return new String[] {
+                StructurizrDslTokens.GROUP_TOKEN,
                 StructurizrDslTokens.DEPLOYMENT_NODE_TOKEN,
                 StructurizrDslTokens.INFRASTRUCTURE_NODE_TOKEN,
                 StructurizrDslTokens.SOFTWARE_SYSTEM_INSTANCE_TOKEN,