+++ docs/changelog.md
@@ -6,6 +6,7 @@
 - Adds support for splitting lines in the DSL source with a backslash character (https://github.com/structurizr/dsl/issues/137).
 - Fixes https://github.com/structurizr/dsl/issues/114 (Parallel sequence behavior in dynamic views).
 - Fixes https://github.com/structurizr/dsl/issues/239 (File context for included files varies based upon how they are included).
+- Adds support for grouping deployment nodes, infrastructure nodes, software system instances, and container instances.
 
 ## 1.29.1 (17th March 2023)